Update README.md
When it comes to selecting the best certified solution for IMAP mail backup, only one name comes to the mind. Mail backup X is the best tool for IMAP email backup and recovery. This tool is well known for its email backup and archiving abilities. You can also use this tool to restore all emails to a live mailbox. Its cross platform compatibility allows it to work with both windows and mac operating systems.